+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Remove prices from messages. 
+- Add ask for cancel message. [#48](https://github.com/DevAlexandreCR/gorda-api/pull/48)
 
 ## [1.1.11 (2023-04-24)](https://github.com/DevAlexandreCR/gorda-api/compare/v1.1.11...v1.1.10)